requirements.txt
numpy>=1.21.0
opencv-python-headless>=4.5.0
Pillow>=8.0.0
streamlit>=1.10.0
torch>=1.9.0  # CPU version
torchvision>=0.10.0  # CPU version
tqdm>=4.62.0
typing-extensions>=4.0.0
scikit-image>=0.19.0  # Added for histogram matching
--extra-index-url https://download.pytorch.org/whl/cpu 